Refactor: moved tracking functions to tracker
